requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.22
pydantic==2.5.3
SQLAlchemy==2.0.23
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
numpy==1.24.3
pandas==2.1.4
tiktoken==0.5.2